requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Web Scraping & APIs
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Data Generation
faker>=19.0.0

# Cloud & Database
google-cloud-bigquery>=3.11.0
pandas-gbq>=0.19.0
db-dtypes>=1.0.0

# Visualization & Dashboard
streamlit>=1.28.0
plotly>=5.0.0

# AI & ML
openai>=1.0.0
langchain>=0.0.300
langchain-openai>=0.0.5
langchain-community>=0.0.10
chromadb>=0.4.0

# Statistics & Analysis
statsmodels>=0.14.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Scheduling & Background Tasks
schedule>=1.2.0